Lab04/requirement.txt
PyQt5
pycryptodome
cryptography
tornado